Move role update to user controller
